+++ R/f_core_constants.R
@@ -13,8 +13,8 @@
 ## | 
 ## |  Contact us for information about our services: info@rpact.com
 ## | 
-## |  File version: $Revision: 5884 $
-## |  Last changed: $Date: 2022-02-25 08:34:20 +0100 (Fr, 25 Feb 2022) $
+## |  File version: $Revision: 5906 $
+## |  Last changed: $Date: 2022-02-26 19:10:21 +0100 (Sa, 26 Feb 2022) $
 ## |  Last changed by: $Author: pahlke $
 ## | 
 
@@ -715,7 +715,8 @@ C_PARAMETER_NAMES <- list(
         designPlan = NULL, 
         stageResults = NULL, 
         analysisResults = NULL,
-        dataset = NULL) {
+        dataset = NULL,
+        designCharacteristics = NULL) {
         
 	parameterNames <- C_PARAMETER_NAMES